Operator: Rename `operators.py` -> `operator.py` to match class name
A the name of a file defining a single class should match the class name.
