+++ devito/operator.py
